refactor: reorganize and rename components and modules
