--- components/certificates/index.tsx
@@ -1,2 +1,0 @@
-export { CertificatesPanel } from './panel'
-export { CertificateCard } from './card'
