+++ components/core/certificates/index.tsx
@@ -1,0 +1,2 @@
+export { CertificatesPanel } from '@/components/core/certificates/panel'
+export { CertificateCard } from '@/components/core/certificates/card'
