+++ components/core/references/index.tsx
@@ -1,0 +1,2 @@
+export { ReferencesPanel } from '@/components/core/references/panel'
+export { ReferenceCard } from '@/components/core/references/card'
