+++ components/core/skills/index.tsx
@@ -1,0 +1,4 @@
+export { SkillBadgesList } from '@/components/core/skills/badge-list'
+export { useSkills } from '@/components/core/skills/useSkills'
+export { SkillBadge } from '@/components/core/skills/badge'
+export type { ToggleSkill, SkillIsActive } from '@/components/core/skills/types'
