+++ components/core/skills/types.ts
