+++ components/site/sections/index.tsx
@@ -1,0 +1,2 @@
+export { CertificationSection } from '@/components/site/sections/certification'
+export { LearnlogSection } from '@/components/site/sections/learnlog'
