--- components/skills/index.tsx
@@ -1,4 +1,0 @@
-export { SkillBadgesList } from './badge-list'
-export { useSkills } from './useSkills'
-export { SkillBadge } from './badge'
-export type { ToggleSkill, SkillIsActive } from '@/components/skills/types'
